Moved stuff from programs to applications, created fixed sidebars and resized dom
