+++ src/components/Header/index.tsx
@@ -1,5 +1,5 @@
 import SettingsIcon from "@mui/icons-material/Settings";
-import { Button } from "@mui/material";
+import { Box, Button } from "@mui/material";
 import Link from "next/link";
 
 import Logo from "@/components/Logo";
@@ -9,35 +9,40 @@ export default async function Header() {
   const session = await getUserSession();
 
   return (
-    <header
-      style={{
-        display: "flex",
-        height: "100px",
-        width: "100%",
-        justifyContent: "space-between",
-        alignItems: "center",
-        paddingInline: "1.5rem",
-        position: "sticky", // Changed from fixed to sticky
-        top: 0, // Required for sticky to work
-        boxShadow: "0 0 10px rgba(0, 0, 0, 0.2)",
-        backgroundColor: "white",
-        zIndex: 999,
-      }}
-    >
-      <Link href="/" style={{ width: "100px", height: "100px" }}>
-        <Logo width={100} height={100} alt="SCF Logo" />
-      </Link>
-      {session && (
-        <Button
-          variant="outlined"
-          color="primary"
-          component={Link}
-          href="/settings"
-          endIcon={<SettingsIcon />}
-        >
-          Settings
-        </Button>
-      )}
-    </header>
+    <>
+      {/* // Shift the dom 100 px down */}
+      <Box style={{ height: "100px" }}></Box>
+      <header
+        style={{
+          display: "flex",
+          height: "100px",
+          width: "100vw",
+          justifyContent: "space-between",
+          alignItems: "center",
+          paddingInline: "1.5rem",
+          position: "fixed",
+          top: 0,
+          left: 0,
+          boxShadow: "0 0 10px rgba(0, 0, 0, 0.2)",
+          backgroundColor: "white",
+          zIndex: 999,
+        }}
+      >
+        <Link href="/" style={{ width: "100px", height: "100px" }}>
+          <Logo width={100} height={100} alt="SCF Logo" />
+        </Link>
+        {session && (
+          <Button
+            variant="outlined"
+            color="primary"
+            component={Link}
+            href="/settings"
+            endIcon={<SettingsIcon />}
+          >
+            Settings
+          </Button>
+        )}
+      </header>
+    </>
   );
 }
